Rename `type` option to `column_type`
Thr reaason for this is that Ecto seems to be handling options
differently when the custom type is used for a `belongs_to` field. For
example if `belongs_to :thing, Thing, type: TypeID` is specified, the
value for `type` will be `TypeIO`, which makes the opts valdiation fail.
Using another name for `type`, makes it possible to both set `type` (The
Ecto type of the field) and `column_type` (The column type in the
database)`.